Code review changes for pull #4282 - Removed hard coded values - Removed uncessary property access calls

+++ Bikewale.Entities/Filters/CustomInputFilters.cs
@@ -7,9 +7,9 @@ namespace Bikewale.Entities.Filters
     /// </summary>
     public class CustomInputFilters
     {
-        public int MinMileage { get; set; }
-        public int MinPrice { get; set; }
-        public int MinDisplacement { get; set; }
+        public ushort MinMileage { get; set; }
+        public long MinPrice { get; set; }
+        public float MinDisplacement { get; set; }
         public uint MakeCategoryId { get; set; }
     }
 }
